dockerized-flask-celery/requirements.txt
flask
celery
python-dotenv
redis
